Initial Checkin
Current status: Has an activity, a backgroundreceiver, and a intentservice. The
activity is in decent shape, close to ready to use. I think the background-
receiver does what it needs to do, but there's no provision for calling it
regularly. The service is in place and being activated properly by the intents,
but it doesn't do anything yet except verify that it can actually get the
preferences out of the shared preferences object.

So the next step is to figure out how sockets work and get to work on making
the actual data exchange, which will involve doing the work in the C# app too.