This attempts to get the auto-sync working right, currently at the interval of one hour. The AutoSync button now creates and cancels a repeating alarm in the alarm manager which triggers the broadcast reciever to operate. A PendingIntent is created at activity creation time to trigger the BroadcastReciever, and the sync now uses it too. The current autosync status is stored as a boolean in the preferences object.
The sync service attempts to use WakeLocks properly too. The WakeLock permission
was added, and the service overrides onStartCommand to try and acquire the
WakeLock before the new thread kicks off and allows the original thread to
return to the alarmmanager and sleep the phone again. The actual function calls
release in a new finally block.

I also split off some of the code from the main sync function into seperate
methods to clean things up a bit, specifically the code to get the messages and
the code to get the secret key from the password. The should mean much less
extra junk in the main method.